update some minor change related menu_group-management.

+++ src/menu-group_management/menu-group_management.service.ts
@@ -4,6 +4,7 @@ import { Repository } from 'typeorm';
 import { MenuGroupManagement } from './entities/menu-group_management.entity';
 import { CreateMenuGroupManagementDto } from './dto/create-menu-group_management.dto';
 import { UpdateMenuGroupManagementDto } from './dto/update-menu-group_management.dto';
+import { MESSAGE } from 'src/common/collection';
 
 @Injectable()
 export class MenuGroupManagementService {
@@ -14,6 +15,13 @@ export class MenuGroupManagementService {
 
   async createMenuGroup(createMenuGroupManagementDto: CreateMenuGroupManagementDto): Promise<MenuGroupManagement> {
     try {
+      const menuGroupExists = await this.findMenuGroupByName(createMenuGroupManagementDto.name);
+      if (menuGroupExists) {
+        throw new HttpException({
+          status: HttpStatus.BAD_REQUEST,
+          error: MESSAGE.WARNING.MENU_GROUP_ALREADY_EXISTS
+        }, HttpStatus.BAD_REQUEST);
+      }
       const newMenuGroup = this.menuGroupManagementRepository.create(createMenuGroupManagementDto);
       return await this.menuGroupManagementRepository.save(newMenuGroup);
     } catch (error) {
@@ -37,7 +45,14 @@ export class MenuGroupManagementService {
 
   async findOneMenuGroup(id: string): Promise<MenuGroupManagement> {
     try {
-      return await this.menuGroupManagementRepository.findOne({ where: { id } });
+      const menuGroup = await this.menuGroupManagementRepository.findOne({ where: { id } });
+      if (!menuGroup) {
+        throw new HttpException({
+          status: HttpStatus.NOT_FOUND,
+          error: MESSAGE.WARNING.MENU_GROUP_NOT_FOUND
+        }, HttpStatus.NOT_FOUND);
+      }
+      return menuGroup;
     } catch (error) {
       throw new HttpException({
         status: error.status ? error.status : HttpStatus.INTERNAL_SERVER_ERROR,
@@ -48,6 +63,13 @@ export class MenuGroupManagementService {
 
   async updateOneMenuGroup(id: string, updateMenuGroupManagementDto: UpdateMenuGroupManagementDto): Promise<MenuGroupManagement> {
     try {
+      const menuGroup = await this.findOneMenuGroup(id);
+      if (!menuGroup) {
+        throw new HttpException({
+          status: HttpStatus.NOT_FOUND,
+          error: MESSAGE.WARNING.MENU_GROUP_NOT_FOUND
+        }, HttpStatus.NOT_FOUND);
+      }
       await this.menuGroupManagementRepository.update(id, updateMenuGroupManagementDto);
       return await this.menuGroupManagementRepository.findOne({ where: { id } });
     } catch (error) {
@@ -60,6 +82,13 @@ export class MenuGroupManagementService {
 
   async removeOneMenuGroup(id: string): Promise<void> {
     try {
+      const menuGroup = await this.findOneMenuGroup(id);
+      if (!menuGroup) {
+        throw new HttpException({
+          status: HttpStatus.NOT_FOUND,
+          error: MESSAGE.WARNING.MENU_GROUP_NOT_FOUND
+        }, HttpStatus.NOT_FOUND);
+      }
       await this.menuGroupManagementRepository.delete(id);
     } catch (error) {
       throw new HttpException({
